+++ src/params/ethBalance.ts
@@ -1,5 +1,5 @@
 import {
-	Opt, ic, match, int16, float32
+	Opt, ic, match, int16, float64
 } from 'azle';
 import decodeUtf8 from 'decode-utf8';
 import { Buffer } from 'buffer';
@@ -9,7 +9,7 @@ import {
 import { formatEther } from '@ethersproject/units';
 
 
-export const ethBalance = async (address: string): Promise<float32> => {
+export const ethBalance = async (address: string): Promise<float64> => {
 
     const response = await managementCanister
 		.http_request({
