+++ src/params/uniswap.ts
@@ -1,5 +1,5 @@
 import {
-	Opt, ic, match, float32
+	Opt, ic, match, float64
 } from 'azle';
 import {
 	managementCanister,
@@ -8,7 +8,7 @@ import decodeUtf8 from 'decode-utf8';
 import { Buffer } from 'buffer';
 
 
-export const uniswap = async (address: string): Promise<float32> => {
+export const uniswap = async (address: string): Promise<float64> => {
     const response = await managementCanister
 		.http_request({
 			url: `https://eth-mainnet.g.alchemy.com/v2/${Buffer.from(process.env.ALCHEMY_ETH!, 'base64').toString('utf-8')}/getNFTs?owner=${address}&withMetadata=false&contractAddresses[]=0xc36442b4a4522e871399cd717abdd847ab11fe88&pageSize=1`,
